Use partial fit to train mnb model

+++ Proyecto 1/Etapa 2/src/lib/prediction_model.py
@@ -2,18 +2,27 @@ from joblib import load
 from cloudpickle import dumps
 
 from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score
-from sklearn.model_selection import train_test_split
 
+import pandas as pd
 import os
 
 class Model:
-    MODELS_DIRNAME = 'assets'
+    ASSETS_DIRNAME = 'assets'
 
     def __init__(self):
         # Load latest model
-        self.models_dir = os.path.join(os.path.dirname(os.path.dirname(__file__)), self.MODELS_DIRNAME)
-        self.models_list = sorted(os.listdir(self.models_dir), reverse=True)
+        self.models_dir = os.path.join(os.path.dirname(os.path.dirname(__file__)), self.ASSETS_DIRNAME)
+        self.models_list = sorted([x for x in os.listdir(self.models_dir) if 'model-' in x], reverse=True)
+        self.validation_loaded = False
         self.__load_model()
+
+    def __load_validation_data(self):
+        # Load validation data
+        validation_data_path = os.path.join(os.path.dirname(os.path.dirname(__file__)), self.ASSETS_DIRNAME, 'validation-data.csv')
+        validation_df = pd.read_csv(validation_data_path)
+        self.x_validation = validation_df['Textos_espanol']
+        self.y_validation = validation_df['sdg']
+        self.validation_loaded = True
 
     def __load_model(self):
         load_path = os.path.join(os.path.dirname(__file__), self.models_dir, self.models_list[0])
@@ -35,19 +44,22 @@ class Model:
         results = map(self.__get_prediction_result, probabilities)
         return list(results)
 
-    def train(self, train_data, test_data):
-        # Split data
-        x_train, x_test, y_train, y_test = train_test_split(train_data, test_data, test_size=0.25, random_state=42)
-        new_model = self.model.fit(x_train, y_train)
+    def train(self, data, labels):
+        if not self.validation_loaded:
+            self.__load_validation_data()
+
+        vectorized_result = self.model.named_steps['preprocessing'].transform(data)
+        self.model.named_steps['clf'].partial_fit(vectorized_result, labels, classes=self.model.named_steps['clf'].classes_)
 
         # Evaluate model
-        y_pred = new_model.predict(x_test)
-        accuracy = accuracy_score(y_test, y_pred)
-        f1 = f1_score(y_test, y_pred, average='weighted')
-        precision = precision_score(y_test, y_pred, average='weighted')
-        recall = recall_score(y_test, y_pred, average='weighted')
+        y_pred = self.model.predict(self.x_validation)
+        accuracy = accuracy_score(self.y_validation, y_pred)
+        f1 = f1_score(self.y_validation, y_pred, average='weighted')
+        precision = precision_score(self.y_validation, y_pred, average='weighted')
+        recall = recall_score(self.y_validation, y_pred, average='weighted')
 
-        self.model = new_model
+        print()
+
         self.save()
 
         return {
@@ -63,6 +75,7 @@ class Model:
         latest_version = self.models_list[0].removeprefix('model-').removesuffix('.pkl')
         new_subversion = int(latest_version.split('.')[1]) + 1
         new_version = latest_version.split('.')[0] + '.' + str(new_subversion)
+        self.model_version = new_version
 
         # Save model
         with open(f"{self.models_dir}/model-{new_version}.pkl", 'wb') as f:
@@ -75,5 +88,7 @@ class Model:
 
     def reset(self):
         for model in self.models_list[:-1]:
+           if 'v1.0' in model:
+               continue
            os.remove(os.path.join(self.models_dir, model))
         self.__load_model()
